split slice plotter into view and presenter classes re #358
